+++ app/components/logo/index.test.js
@@ -1,0 +1,12 @@
+import {mount} from 'enzyme'
+import React from 'react'
+import 'ignore-styles'
+
+import Logo from './index'
+
+/* eslint-disable newline-per-chained-call */
+
+test('Logo renders without errors', () => {
+    const wrapper = mount(<Logo />)
+    expect(wrapper.length).toBe(1)
+})