+++ app/containers/login/actions.js
@@ -1,0 +1,3 @@
+import {createAction} from '../../utils/utils'
+
+export const myAction = createAction('Login page loaded')