+++ app/loader-routes.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 // GENERATED FILE DO NOT EDIT
-const routes = [/.*/]
+const routes = [/^\/\/?$/,/^\/customer\/account\/login\/?$/,/^\/.*?\.html\/?$/]
 export default routes