+++ app/loader.js
@@ -1,5 +1,6 @@
-import {getAssetUrl, loadAsset} from 'progressive-web-sdk/dist/asset-utils'
+import {getAssetUrl, loadAsset, initCacheManifest} from 'progressive-web-sdk/dist/asset-utils'
 import {displayPreloader} from 'progressive-web-sdk/dist/preloader'
+import cacheHashManifest from '../tmp/loader-cache-hash-manifest.json'
 
 window.Progressive = {}
 
@@ -9,10 +10,12 @@ const isReactRoute = () => {
     return ReactRegexes.some((regex) => regex.test(window.location.pathname))
 }
 
+initCacheManifest(cacheHashManifest)
+
 const CAPTURING_CDN = '//cdn.mobify.com/capturejs/capture-latest.min.js'
 
 import preloadHTML from 'raw!./preloader/preload.html'
-import preloadCSS from 'raw!./preloader/preload.css'
+import preloadCSS from 'css?minimize!./preloader/preload.css'
 import preloadJS from 'raw!./preloader/preload.js' // eslint-disable-line import/default
 
 if (isReactRoute()) {