+++ webpack/base.loader.js
@@ -2,7 +2,7 @@
 /* eslint-env node */
 
 const path = require('path')
-const CACHE_MANIFEST_NAME = 'loader-cache-hash-manifest.json'
+const baseCommon = require('./base.common')
 
 module.exports = {
     devtool: 'cheap-source-map',
@@ -17,9 +17,6 @@ module.exports = {
         root: path.join(process.cwd(), 'node_modules')
     },
     resolve: {
-        alias: {
-            cacheHashManifest: path.resolve(process.cwd(), 'tmp', CACHE_MANIFEST_NAME)
-        },
         extensions: ['', '.js', '.jsx', '.json']
     },
     module: {
@@ -31,9 +28,15 @@ module.exports = {
                 cacheDirectory: `${__dirname}/tmp`
             },
             {
+                test: /\.css?$/,
+                exclude: /node_modules/,
+                loader: 'postcss',
+            },
+            {
                 test: /\.json$/,
                 loader: 'json'
             }
         ],
-    }
+    },
+    postcss: baseCommon.postcss
 }