+++ CHANGELOG.md
@@ -1,8 +1,14 @@
-## 0.0.4 (August 31)
+## 0.1.0 (September 14, 2016)
+- Add PostCSS (Autoprefixer) and CSS minification
+- Move Styleguidist's webpack settings to the `/webpack` directory
+- Use SDK version 0.4.3 (test framework AVA replaced with Jest, Analytics utils)
+- Project utils.js now contains makeRequest method
+
+## 0.0.4 (August 31, 2016)
 - Add configuration for building in specific cache manifest files into the main
   and loader build files
 
-## 0.0.3 (August 26)
+## 0.0.3 (August 26, 2016)
 - use SDK version 0.1.1
 
 ## 0.0.2